+++ aviasales2/src/main/java/com/example/aviasales2/service/Validation.java
@@ -1,6 +1,6 @@
 package com.example.aviasales2.service;
 
 public interface Validation {
-    Double checkSum(long userId,long tripId,int count);
-    Double checkSeats(long tripId, int count);
+    int checkSum(long userId,long tripId,int count);
+    int checkSeats(long tripId, int count);
 }
